Fixed stuff like database, chat bubble

+++ models/numbers.js
@@ -1,7 +1,7 @@
 module.exports = function (sequelize, DataTypes) {
 	var Numbers = sequelize.define("Numbers", {
 		name: DataTypes.STRING,
-		number: DataTypes.INTEGER
+		number: DataTypes.BIGINT
 	});
 	return Numbers;
 };
